business_server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pymongo==4.13.0
requests==2.31.0
aiohttp==3.11.18 
python-dotenv==1.1.0
httpx==0.28.1